backend/kiwoom_bridge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
supabase==2.0.2
pydantic==2.5.2
python-multipart==0.0.6
httpx>=0.24.0,<0.26.0
pandas==2.1.4
numpy==1.24.3
